Add optional taxa functionality, predicted table, predicted legend with filter functionality, taxon details window and first draft for visualized graph
Commit adds a bunch of stuff like: functionality for using the optional taxa field on home to use the /predict endpoint for specific taxa. Predicted table with a legend for the prediction scores, this legend also allows to filter in the table. A taxon view for view details of a taxon like the biological hierarchy when you select one from one of the tables. First draft version of the visualized graph.

Add;
- Optional taxa for /predict endpoint
- Predict table
- Predict legend
- Filter functionality for predict table based on prediction score
- Taxon details section
- Draft for interlinked taxa graph