Fix Relation Graph in Chrome
Commit fixes the bug that caused the relation graph in Chrome to not extend upon existing nodes. The issue originated in the query form component because it did not update the interaction type variable. This was caused by Chrome not supporting onclick events for select options, instead use onchange on the select element.